pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aion-framework"
version = "1.0.0"
description = "The Durable Application Framework for Agentic AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hatchet-sdk>=0.36.0",
    "pydantic-ai>=0.0.14",
    "openai>=1.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "lancedb>=0.17.0",
    "pandas>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[project.scripts]
aion = "aion.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
